add lib structure and refactor code
