+++ lib/constants.js
@@ -14,4 +14,8 @@ const categories = {
   BREAKING: 'BREAKING'
 };
 
-module.exports = { status, categories, argsNames };
+const { success, warning, error } = require('log-symbols');
+
+const criticalitySymbols = { BREAKING: error, DANGEROUS: warning, NON_BREAKING: success };
+
+module.exports = { status, categories, argsNames, criticalitySymbols };
