+++ lib/helper/schemas.js
@@ -3,9 +3,7 @@ const { logger } = require('express-wolox-logger');
 const { buildSchema } = require('graphql');
 const { importSchema } = require('graphql-import');
 const { diff } = require('@graphql-inspector/core');
-const { success, warning, error } = require('log-symbols');
-
-const criticalitySymbols = { BREAKING: error, DANGEROUS: warning, NON_BREAKING: success };
+const { criticalitySymbols } = require('../constants');
 
 const mergeSchemas = (schemaToWrite, schemaToRead) => {
   logger.info('Merging schemas');
