+++ lib/resolvers/inspector.js
@@ -20,7 +20,7 @@ const resolveInspection = (changes, oldSchemaFile, newSchema) => {
       process.exit(status.ERROR_EXIT_STATUS);
     }
   } else {
-    logger.info(success, chalk.green.bold('No breaking changes detected in schema'));
+    logger.info(success, chalk.green.bold('No breaking changes detected'));
     mergeSchemas(oldSchemaFile, newSchema);
     process.exit(status.NO_ERROR_EXIT_STATUS);
   }
